requirements.txt
MetaTrader5==5.0.45
pandas==2.2.2
numpy==1.26.4
python-telegram-bot==21.5
mplfinance==0.12.10b0
matplotlib==3.9.2
ipython==8.26.0